+++ src/app/pages/home/calendar-day-header/calendar-day-header.component.ts
@@ -1,7 +1,10 @@
 import { Component, Input, Output, EventEmitter } from '@angular/core';
 import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
 
 import { CalendarView } from 'angular-calendar';
+
+import { EventsSearchDialogComponent } from '@features/events/events-search-dialog/events-search-dialog.component';
 
 @Component({
   selector: 'app-calendar-day-header',
@@ -20,9 +23,30 @@ export class CalendarDayHeaderComponent {
   CalendarView = CalendarView;
 
 
-  constructor(private router: Router) {}
+  constructor(
+    public dialog: MatDialog,
+    private router: Router
+  ) {}
 
   gotoCalendar() {
     this.router.navigate([`calendario`]);
+  }
+
+  gotoEventSearch() {
+    console.log(`gotoEventSearch!`);
+
+    const dialogRef = this.dialog.open(EventsSearchDialogComponent);
+
+    // dialogRef.afterClosed().subscribe((eventDialog: IEvent) => {
+    //   if ( eventDialog ) {
+    //     this.event.name = eventDialog.name;
+    //     this.event.description = eventDialog.description;
+    //     this.event.categories = eventDialog.categories;
+    //     this.eventSrv.updateEvent(this.event, AuditType.UPDATED_INFO);
+    //   } else {
+    //     this.utilsSrv.swalFire(SwalMessage.NO_CHANGES);
+    //   }
+    // });
+
   }
 }